+++ CHANGES.md
@@ -5,6 +5,7 @@ Changes in 3.1.2
 ----------------
 - Fix Assignment view: do not exit so Warehouse('footer') is called in StudentAssignments.php
 - Remove Used for Attendance column, unused in Periods.php
+- Add Student photo Tip message in MakeReferral.php & Referrals.php
 
 Changes in 3.1.1
 ----------------